feat: Update WPdash Child version 6.1 (#6)
* feat: Update WPdash Child version 6.1

The commit removes outdated and no longer necessary files from the phpseclib vendor directory. This includes a number of composer.json, README.md, and LICENSE.txt files, amongst others, for different libraries. This clean-up may enhance maintainability and reduce potential confusion about which files and dependencies are currently relevant.

* feat: Update 'Tested up to' version in readme

The 'Tested up to' version in the readme file has been updated from 6.4.2 to 6.5.3. This reflects the ongoing compatibility testing with newer versions of the software.